+++ src/routes/package.ts
@@ -6,7 +6,7 @@ import type {Src_Json} from '@ryanatkn/gro/src_json.js';
 export const package_json = {
 	name: '@ryanatkn/fuz_code',
 	version: '0.20.0',
-	description: 'syntax styling utilities and components for Svelte, SvelteKit, TypeScript',
+	description: 'syntax styling utilities and components for TypeScript, Svelte, SvelteKit',
 	glyph: '🎨',
 	logo: 'logo.svg',
 	logo_alt: 'a friendly pink spider facing you',